[modularize] Subclassing `ImportThread`.
